Done with right swap
In this from frontend we will get logged in user id and right swiped user id. first check if both of the entries already exist in the table then convert make it a match (make the boolean variable true) else add that in the table and keep the value as false.